+++ src/Phan/BlockAnalysisVisitor.php
@@ -709,6 +709,14 @@ class BlockAnalysisVisitor extends AnalysisVisitor
                 BlockExitStatusChecker::willUnconditionallyProceed($stmts_node)
             ))->checkRedundantOrImpossibleTruthyCondition($condition_node, $context, null, false);
         }
+
+        // Give plugins a chance to analyze the loop condition now
+        ConfigPluginSet::instance()->analyzeLoopBeforeBody(
+            $this->code_base,
+            $context,
+            $node
+        );
+
         $context = $this->analyzeAndGetUpdatedContext(
             $context->withScope(
                 new BranchScope($context->getScope())
@@ -849,6 +857,13 @@ class BlockAnalysisVisitor extends AnalysisVisitor
             ))->checkRedundantOrImpossibleTruthyCondition($condition_node, $context, null, false);
         }
 
+        // Give plugins a chance to analyze the loop condition now
+        ConfigPluginSet::instance()->analyzeLoopBeforeBody(
+            $this->code_base,
+            $context,
+            $node
+        );
+
         $context = $this->analyzeAndGetUpdatedContext(
             $context->withScope(
                 new BranchScope($context->getScope())
@@ -962,6 +977,14 @@ class BlockAnalysisVisitor extends AnalysisVisitor
         if ($key_node instanceof Node) {
             $inner_context = $this->analyzeAndGetUpdatedContext($inner_context, $node, $key_node);
         }
+
+        // Give plugins a chance to analyze the loop condition now
+        ConfigPluginSet::instance()->analyzeLoopBeforeBody(
+            $code_base,
+            $context,
+            $node
+        );
+
         $stmts_node = $node->children['stmts'];
         if ($stmts_node instanceof Node) {
             $inner_context = $this->analyzeAndGetUpdatedContext($inner_context, $node, $stmts_node);